+++ src/App.css
@@ -114,7 +114,6 @@ body {
     width: 310px;
     height: 460px;
     margin: 1.5rem;
-    position: relative;
     border-radius: 12px;
     overflow: hidden;
     border: none;
@@ -122,19 +121,19 @@ body {
     box-shadow: 0px 13px 10px -7px rgba(0, 0, 0, 0.1);
 }
 
-.movie div:nth-of-type(1) {
-    position: absolute;
+.movie .react-card-front div:nth-of-type(1) {
     padding: 16px;
     width: 100%;
     opacity: 0;
     top: 0;
     color: #f9d3b4;
+    perspective: 300px;
 }
 
-.movie:hover {
+/* .movie:hover {
     box-shadow: 0px 30px 18px -8px rgba(0, 0, 0, 0.1);
     transform: scale(1.05, 1.05);
-}
+} */
 
 .movie div:nth-of-type(2) {
     width: 100%;
@@ -144,6 +143,12 @@ body {
 .movie div:nth-of-type(2) img {
     height: 100%;
     width: 100%;
+}
+
+.react-card-back div:nth-of-type(1) {
+    height: 460px;
+    width: 310px;
+    background-color: #f0f0f0;
 }
 
 .movie div:nth-of-type(3) {
@@ -173,22 +178,22 @@ body {
     color: #f9d3b4;
 }
 
-.movie:hover div:nth-of-type(2) {
+/* .movie:hover div:nth-of-type(2) {
     height: 100%;
     opacity: 0.3;
-}
+}  */
 
-.movie:hover div:nth-of-type(3) {
+/* .movie:hover div:nth-of-type(3) {
     background-image: none;
     -webkit-transition-duration: 600ms;
     transition-duration: 200ms;
     transform: translateY(10%);
     transform: translateX(-1%);
-}
+}  */
 
-.movie:hover div:nth-of-type(1) {
+/* .movie:hover div:nth-of-type(1) {
     opacity: 1;
-}
+}  */
 
 @media screen and (max-width: 1200px) {
     .app {
